requirements.txt
streamlit
pandas
plotly
requests
fpdf2
kaleido
python-dotenv
matplotlib